requirements.txt
discord.py
requests
python-dotenv
asyncpg
asyncpraw
flask